point du 10/04/2023

+++ Common Programming Concepts/Basic Data Types/Numeric Test/src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
-    let fahrenheit_degree = 451/*TODO*/;
-    let celsius_degree = (fahrenheit_degree - 32/*TODO*/) * 0.5556;
+    let fahrenheit_degree = 451 as f32;
+    let celsius_degree = (fahrenheit_degree - 32 as f32) * 0.5556;
     println!("{}°F is {:.1}°C", fahrenheit_degree, celsius_degree);
 }
